+++ src/main/java/com/penekhun/ctfjserver/User/Controller/ProblemController.java
@@ -65,7 +65,7 @@ public class ProblemController {
             @ApiResponse(responseCode = "501", description = "어드민은 문제를 풀 수 없음", ref = "#/components/responses/ErrorCode.ONLY_ACCESS_USER"),
             @ApiResponse(responseCode = "403", description = "이미 맞춤", ref = "#/components/responses/ErrorCode.ALREADY_CORRECT"),
             @ApiResponse(responseCode = "404", description = "잘못된 플래그(오답)", ref = "#/components/responses/ErrorCode.INCORRECT_FLAG")})
-    public ResponseEntity<String> authProblemMapping(@CurrentUserParameter Account account, @PathVariable @Validated @NotNull Integer problemId, @Valid ProblemDto.Req.Auth auth){
+    public ResponseEntity<String> authProblemMapping(@CurrentUserParameter Account account, @PathVariable @Validated @NotNull Long problemId, @Valid ProblemDto.Req.Auth auth){
         boolean isCorrect = problemService.authProblem(account, problemId, auth);
         logService.authProblemLog(problemId, auth.getFlag(), isCorrect);
         if (isCorrect)
